refactor: separate unit tests from integration tests in pkg/matchers
- Created matchers_integration_test.go for registry-based tests
- Moved TestCreateMatcher and TestValidateMatcher to integration tests
- Kept pure logic tests (sorting, filtering, merging) as unit tests
- Unit tests now only test pure functions without registry dependencies
